az-list ID, manifest, and search.php
az-list ID - generates an HTML page for displaying an anime list, fetching data from an API, and providing navigation through the list using pagination links.

Manifest.json - is a JSON object that defines the basic configuration settings for the AniKatsu Progressive Web App, including its name, icons, start URL, colors, and description.

search.php - is a PHP script that generates an HTML page that displays search results for anime based on a specified keyword. It retrieves data from an API, formats it, and presents it to the user.