movie.js, plyr.min.js and ongoing page
movie.js - JavaScript code for a web page or application related to movie streaming and commenting.

ongoing.php - This PHP code is used to generate an HTML page that displays ongoing anime on a website. It retrieves data from an API and dynamically populates the page with information such as anime titles, images, and status. The page also includes pagination for navigating through the list of ongoing anime. This code can be used as part of a larger website or application that focuses on anime streaming and provides a user-friendly interface for browsing and watching ongoing anime series.