.htaccess, _config, sitemaps
allanime - PHP script generates an XML sitemap for a website by fetching anime data from an API and constructing URLs based on the anime IDs. The sitemap includes <url> elements with information about the URL, change frequency, and priority for search engine indexing.

sitemap.php - XML Sitemap Generator for Website: This PHP script generates an XML sitemap for a website. It includes a configuration file and outputs the sitemap in the proper XML format. The script dynamically generates URLs for the website's homepage, various categories, sub-categories, and seasonal anime pages. Each URL includes a change frequency and priority tag for search engine optimization. The script uses server variables to construct the URLs and sets the appropriate content type header. It is designed to be easily customizable and can be integrated into existing websites.